refactor: 폴더 구조 및 import 정리
