+++ src/app/(route)/saves/[category]/_component/index.ts
@@ -1,4 +1,5 @@
-import SaveFolderItemList from './SaveItemList'
-import MoveFolderModal from './MoveFolderModal'
+import SaveItem from './SaveItem'
+import SaveItemList from './SaveItemList'
+import SaveFolderHeader from './SaveFolderHeader'
 
-export { SaveFolderItemList, MoveFolderModal }
+export { SaveItemList, SaveItem, SaveFolderHeader }
