+++ src/app/_types/save.type.ts
@@ -1,0 +1,5 @@
+export const enum SavePageMode {
+  DEFAULT = 0,
+  CHANGE_NAME = 1,
+  EDIT_LIST = 2,
+}
